feat: restructure resolving and migrating
